+++ CHANGELOG.md
@@ -23,6 +23,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 - `RDF.XSD.Double` and `RDF.XSD.Float` literals created from Elixir floats and integers
   are now interpreted to be in the canonical lexical form
+- the Turtle encoder now uses the uncanonical form of a literal in order to not lose
+  the lexical form of a literal
 - `RDF.BlankNode.new/1` ignores the prefix `"_:"` in a given blank node name
 
 ### Fixed